Initial sketch of the event model
No relations for now. Also I believe that the data model can be a bit
more granular if we are to extract the city and the country in separate
tables. It would be nice if even have an address an its in it's own
table, along with longitude and latitude information.

However I wouldn't like to go that deep into the granular approach, as I
would love to not have to deal with events and duplicate the effort that
railsgirls.com is doing.